python-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
httpx==0.26.0
pydantic==2.6.0
python-dotenv==1.0.1
numpy==1.26.3
scipy==1.11.4

